moved json parsing out of AbstractCRUDController
